pyproject.toml
[project]
name = "adk-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.17.0",
    "google-generativeai>=0.8.5",
    "litellm>=1.79.1",
    "psutils>=3.3.14",
    "python-dotenv>=1.2.1",
    "yfinance>=0.2.66",
]
